requirements.txt
selenium==3.141.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
git+https://github.com/GPelayo/binge-companion-models@prod#egg=binge_companion_models